schemas

+++ backend/main.py
@@ -1,21 +1,19 @@
-from fastapi import FastAPI,Depends,status,HTTPException,Query, Request
-from pydantic import BaseModel
-from typing import Annotated
+from fastapi import FastAPI,Depends,status,HTTPException,Query
+from typing import Annotated,Dict
 import models
+import schemas
 from database import engine,SessionLocal
 from sqlalchemy.orm import Session
-from datetime import date
-from typing import Optional,List
+from typing import Optional
 from fastapi.middleware.cors import CORSMiddleware
 import uuid
+from fastapi.security import OAuth2PasswordBearer
 import requests
-
 
 app=FastAPI()
 models.Base.metadata.create_all(bind=engine)
 
-app = FastAPI()
-
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 origins = [
     "http://localhost:5173",  # React development server
@@ -29,81 +27,6 @@ app.add_middleware(
     allow_headers=["*"],  # Allows all headers
 )
 
-class EmployeeDetailsCreate(BaseModel):
-    employee_id: str
-    name: str
-    email_id: str
-    picture: str
-    role: str
-
-class ProjectDetailsCreate(BaseModel):
-    title: str
-    description: str
-    startdate: date
-    enddate: date
-    project_status: str
-    class Config:
-        orm_mode = True 
-
-class CreateProject(BaseModel):
-    project: ProjectDetailsCreate
-    employees: List[str]
-
-class ProjectUpdate(BaseModel):
-    title: Optional[str] = None
-    description: Optional[str] = None
-    startdate: Optional[date] = None
-    enddate: Optional[date] = None
-    project_status: Optional[str] = None
-
-class UpdateProject(BaseModel):
-    project: ProjectUpdate
-    employees: Optional[List[str]] = None
-
-class TaskDetailsCreate(BaseModel):
-    project_id: str
-    title: str
-    description: str
-    startdate: date
-    enddate: date
-    task_status: str
-    class Config:
-        orm_mode = True 
-
-class Createtask(BaseModel):
-    task: TaskDetailsCreate
-    employees: List[str]
-
-class TaskUpdate(BaseModel):
-    title: Optional[str] = None
-    description: Optional[str] = None
-    startdate: Optional[date] = None
-    enddate: Optional[date] = None
-    task_status: Optional[str] = None
-
-class UpdateTask(BaseModel):
-    task: TaskUpdate
-    employees: Optional[List[str]] = None
-
-{
-# class TaskDetailsResponse(BaseModel):
-#     employee_id: int
-#     project_id: str
-#     task_id: str
-#     title: str
-    
-#     class Config:
-#         orm_mode = True
-}
-
-class EmployeeProjectDetails(BaseModel):
-    project_id: str
-    employee_id: str
-    role: str
-
-class GoogleLoginRequest(BaseModel):
-    token: str
-
 def get_db():
     db = SessionLocal()
     try:
@@ -113,10 +36,26 @@ def get_db():
 
 db_dependency = Annotated[Session,Depends(get_db)]
 
+def verify_google_token(db:db_dependency,token: str = Depends(oauth2_scheme)) -> Dict:
+    
+    url = f"https://www.googleapis.com/oauth2/v3/tokeninfo?access_token={token}"
+    response = requests.get(url)
 
+    if response.status_code != 200:
+        raise HTTPException(status_code=401, detail="Invalid Google access token")
+    
+    user_info = response.json()
+    
+    db_role = db.query(models.EmployeeDetails.role).filter(models.EmployeeDetails.employee_id == user_info.get('sub')).first()
+    user_info["role"] = db_role[0]
+
+    return user_info 
 
 @app.post("/project",status_code=status.HTTP_201_CREATED)
-async def create_project(project_details:CreateProject,db:db_dependency):
+async def create_project(project_details:schemas.CreateProject,db:db_dependency, user: dict = Depends(verify_google_token)):
+    if user["role"] == "user":
+        return "Your are not an admin"
+    
     project = project_details.project
     project_id = str(uuid.uuid4())
     db_project=models.ProjectDetails(project_id=project_id,**project.model_dump())
@@ -130,9 +69,8 @@ async def create_project(project_details:CreateProject,db:db_dependency):
     print("end")
 
 
-
 @app.get('/project/{id}', status_code=status.HTTP_200_OK)
-async def projectid(db:db_dependency,id:str):
+async def projectid(db:db_dependency,id:str, user: dict = Depends(verify_google_token)):
     project = db.query(models.ProjectDetails).filter(models.ProjectDetails.project_id == id).first()
     if project is None:
         return HTTPException(status_code=404, detail="No project's found")
@@ -140,21 +78,22 @@ async def projectid(db:db_dependency,id:str):
 
 
 @app.get('/project', status_code=status.HTTP_200_OK)
-async def get_project(db: db_dependency, employee_id: Optional[str] = Query(None)):
-
+async def get_project(db: db_dependency, employee_id: Optional[str] = Query(None), user: dict = Depends(verify_google_token)):
     projects=[]
+
     if employee_id:
-        print("employee_id")
         employee = db.query(models.EmployeeDetails).filter(models.EmployeeDetails.employee_id == employee_id).first()
+        
         if not employee:
             raise HTTPException(status_code=404, detail="Employee not found")
         project_list = db.query(models.EmployeeProjectsDetails).filter(models.EmployeeProjectsDetails.employee_id == employee_id).all()
-        print(project_list)
+        
         for pro in project_list:
             print(pro)
             project = db.query(models.ProjectDetails).filter(models.ProjectDetails.project_id == pro.project_id).first()
             print(project)
             projects.append(project)
+
     else:
         project=db.query(models.ProjectDetails).all()
         return project
@@ -163,9 +102,10 @@ async def get_project(db: db_dependency, employee_id: Optional[str] = Query(None
 
 
 @app.delete('/project/{project_id}',status_code=status.HTTP_200_OK)
-async def delete_project(project_id:str,db:db_dependency):
-    print("project_delete")
+async def delete_project(project_id:str,db:db_dependency,user: dict = Depends(verify_google_token)):
 
+    if user["role"] == "user":
+        return "Your are not an admin"
     tasks = db.query(models.TaskDetails).filter(models.TaskDetails.project_id == project_id)
 
     for task in tasks:
@@ -174,9 +114,11 @@ async def delete_project(project_id:str,db:db_dependency):
 
     project = db.query(models.ProjectDetails).filter(models.ProjectDetails.project_id == project_id).first()
     projectemp = db.query(models.EmployeeProjectsDetails).filter(models.EmployeeProjectsDetails.project_id == project_id).all()
+    
     for pro in projectemp:
         db.delete(pro)
         db.commit()
+    
     if project is None:
         return HTTPException(status_code=404, detail='Project Not found')
     db.delete(project)
@@ -184,7 +126,10 @@ async def delete_project(project_id:str,db:db_dependency):
 
 
 @app.put('/project/{project_id}', status_code=status.HTTP_200_OK)
-async def update_project(project_id: str, project_update: UpdateProject, db: db_dependency):
+async def update_project(project_id: str, project_update: schemas.UpdateProject, db: db_dependency,user: dict = Depends(verify_google_token)):
+    
+    if user["role"] == "user":
+        return "Your are not an admin"
     project = db.query(models.ProjectDetails).filter(models.ProjectDetails.project_id == project_id).first()
     if project is None:
         raise HTTPException(status_code=404, detail="Project not found")
@@ -207,43 +152,44 @@ async def update_project(project_id: str, project_update: UpdateProject, db: db_
             db.commit()
 
 
-
-
 # ________________________________________________________________________________________________________
 
 
-
-
-
 @app.post('/task',status_code=status.HTTP_201_CREATED)
-async def create_task(task_details:Createtask,db:db_dependency):
-    db_project = db.query(models.ProjectDetails).filter(models.ProjectDetails.project_id == task_details.task.project_id).first()
-    if not db_project:
-        raise HTTPException(status_code=404, detail="project not found")
+async def create_task(task_details:schemas.Createtask,db:db_dependency, user: dict = Depends(verify_google_token)):
 
-    task=task_details.task
-    task_id=str(uuid.uuid4())
-    db_task=models.TaskDetails(task_id=task_id,**task.model_dump())
-    db.add(db_task)
-    db.commit()
-    id_project=task_details.task.project_id
+    manager_ids = await get_projects(db,task_details.task.project_id)
 
-    for employee in task_details.employees:
-        print(employee)
-        db_employeetask = models.EmployeeTasksDetails(task_id=task_id,employee_id=employee)
-        db.add(db_employeetask)
+    if user.get('sub') in manager_ids and user["role"] == "Admin":
+        db_project = db.query(models.ProjectDetails).filter(models.ProjectDetails.project_id == task_details.task.project_id).first()
+        if not db_project:
+            raise HTTPException(status_code=404, detail="project not found")
+        
+        task=task_details.task
+        task_id=str(uuid.uuid4())
+        db_task=models.TaskDetails(task_id=task_id,**task.model_dump())
+        db.add(db_task)
         db.commit()
-    
-    for employee in task_details.employees:
-        print(employee)
-        db_emp_project = db.query(models.EmployeeProjectsDetails).filter(models.EmployeeProjectsDetails.employee_id == employee).filter(models.EmployeeProjectsDetails.project_id == id_project).first()
-        print(db_emp_project)
-        if db_emp_project:
-            print("Alreay exists")
-        else:
-            db_employeeproject = models.EmployeeProjectsDetails(project_id=task.project_id,employee_id=employee,role="employee")
-            db.add(db_employeeproject)
+        id_project=task_details.task.project_id
+
+        for employee in task_details.employees:
+            print(employee)
+            db_employeetask = models.EmployeeTasksDetails(task_id=task_id,employee_id=employee)
+            db.add(db_employeetask)
             db.commit()
+        
+        for employee in task_details.employees:
+            print(employee)
+            db_emp_project = db.query(models.EmployeeProjectsDetails).filter(models.EmployeeProjectsDetails.employee_id == employee).filter(models.EmployeeProjectsDetails.project_id == id_project).first()
+            print(db_emp_project)
+            if db_emp_project:
+                print("Already exists")
+            else:
+                db_employeeproject = models.EmployeeProjectsDetails(project_id=task.project_id,employee_id=employee,role="employee")
+                db.add(db_employeeproject)
+                db.commit()
+    else:
+        return "Your are not a manager or admin to create a Task" 
 
 
 @app.get('/task/{id}', status_code=status.HTTP_200_OK)
@@ -257,6 +203,7 @@ async def all_task(db:db_dependency,id:str):
 @app.get('/task', status_code=status.HTTP_200_OK)
 async def get_task(db: db_dependency,project_id: Optional[str] = Query(None), employee_id: Optional[str] = Query(None)):
     tasks=[]
+
     if project_id and employee_id:
         all_tasks = db.query(models.TaskDetails).filter(models.TaskDetails.project_id == project_id)
         tasks_with_employees = all_tasks.join(
@@ -269,12 +216,15 @@ async def get_task(db: db_dependency,project_id: Optional[str] = Query(None), em
         tasks = db.query(models.TaskDetails).filter(models.TaskDetails.project_id == project_id).all()
     elif employee_id:
         tasks_emp = db.query(models.EmployeeTasksDetails).filter(models.EmployeeTasksDetails.employee_id == employee_id).all()
-        for task_emp in tasks_emp:
-            task = db.query(models.TaskDetails).filter(models.TaskDetails.task_id == task_emp.task_id).first()
-            tasks.append(task)
+        if tasks_emp:
+            for task_emp in tasks_emp:
+                task = db.query(models.TaskDetails).filter(models.TaskDetails.task_id == task_emp.task_id).first()
+                tasks.append(task)
+        else:
+            raise HTTPException(status_code=404, detail="No Employee Found")
     else:
         tasks=db.query(models.TaskDetails).all()
-    print("sjvh")
+    
     if not tasks:
         raise HTTPException(status_code=404, detail="Task not found")
     
@@ -282,38 +232,54 @@ async def get_task(db: db_dependency,project_id: Optional[str] = Query(None), em
 
 
 @app.delete('/task/{task_id}',status_code=status.HTTP_200_OK)
-async def delete_task(task_id:str,db:db_dependency):
-    print("delete_task")
-    task = db.query(models.TaskDetails).filter(models.TaskDetails.task_id == task_id).first()
-    task_employee = db.query(models.EmployeeTasksDetails).filter(models.EmployeeTasksDetails.task_id == task_id).all()
-    if task is None:
-        return HTTPException(status_code=404, detail='task Not found')
-    for t in task_employee:
-        db.delete(t)
-        db.commit()
-    db.delete(task)
-    db.commit()
-
-@app.put('/task/{task_id}',status_code=status.HTTP_200_OK)
-async def taskid(task_id:str,task_update:UpdateTask,db:db_dependency):
-    task = db.query(models.TaskDetails).filter(models.TaskDetails.task_id == task_id).first()
-    if task is None:
-        return HTTPException(status_code=404, detail='task Not found')
-    for key, value in task_update.task.model_dump(exclude_unset=True).items():
-        setattr(task, key, value) 
-    db.commit()
-
-    taskemployee = db.query(models.EmployeeTasksDetails).filter(models.EmployeeTasksDetails.task_id == task_id).all()
+async def delete_task(task_id:str,db:db_dependency,user: dict = Depends(verify_google_token)):
     
-    for task in taskemployee:
+    project_id = db.query(models.TaskDetails.project_id).filter(models.TaskDetails.task_id == task_id).first()
+
+    manager_ids = await get_projects(db,project_id[0])
+
+    if user.get('sub') in manager_ids and user["role"] == "Admin":
+        task = db.query(models.TaskDetails).filter(models.TaskDetails.task_id == task_id).first()
+        task_employee = db.query(models.EmployeeTasksDetails).filter(models.EmployeeTasksDetails.task_id == task_id).all()
+        if task is None:
+            return HTTPException(status_code=404, detail='task Not found')
+        for t in task_employee:
+            db.delete(t)
+            db.commit()
         db.delete(task)
         db.commit()
+    else:
+        return "Your neither an admin nor Manager"
 
-    if task_update.employees:
-        for employee in task_update.employees:
-            db_projectemployee=models.EmployeeTasksDetails(task_id=task_id,employee_id=employee)
-            db.add(db_projectemployee)
+
+@app.put('/task/{task_id}',status_code=status.HTTP_200_OK)
+async def taskid(task_id:str,task_update:schemas.UpdateTask,db:db_dependency,user: dict = Depends(verify_google_token)):
+
+    project_id = db.query(models.TaskDetails.project_id).filter(models.TaskDetails.task_id == task_id).first()
+
+    manager_ids = await get_projects(db,project_id[0])
+
+    if user.get('sub') in manager_ids and user["role"] == "Admin":
+        task = db.query(models.TaskDetails).filter(models.TaskDetails.task_id == task_id).first()
+        if task is None:
+            return HTTPException(status_code=404, detail='task Not found')
+        for key, value in task_update.task.model_dump(exclude_unset=True).items():
+            setattr(task, key, value) 
+        db.commit()
+
+        taskemployee = db.query(models.EmployeeTasksDetails).filter(models.EmployeeTasksDetails.task_id == task_id).all()
+        
+        for task in taskemployee:
+            db.delete(task)
             db.commit()
+
+        if task_update.employees:
+            for employee in task_update.employees:
+                db_projectemployee=models.EmployeeTasksDetails(task_id=task_id,employee_id=employee)
+                db.add(db_projectemployee)
+                db.commit()
+    else:
+        return "Your neither an admin nor Manager"
 
 
 
@@ -348,7 +314,7 @@ async def get_projects(db:db_dependency, project_id: Optional[str] = Query(None)
         ).distinct().all()
         print(result)
 
-        return [emp.employee_id for emp in result]
+        return result
     emp_pro = db.query(models.EmployeeProjectsDetails).all()
     return emp_pro
 
@@ -359,7 +325,7 @@ async def get_tasks(db:db_dependency, task_id: Optional[str] = Query(None)):
             models.EmployeeTasksDetails.task_id == task_id
         ).distinct().all()
 
-        return [emp.employee_id for emp in result]
+        return result
 
     emp_task = db.query(models.EmployeeTasksDetails).all()
     return emp_task
@@ -368,7 +334,7 @@ async def get_tasks(db:db_dependency, task_id: Optional[str] = Query(None)):
 
 
 @app.post("/auth/google")
-async def authenticate_google_user(request: GoogleLoginRequest,db:db_dependency):
+async def authenticate_google_user(request: schemas.GoogleLoginRequest,db:db_dependency):
     access_token = request.token
     url = f"https://www.googleapis.com/oauth2/v3/userinfo?access_token={access_token}"
     response = requests.get(url)
@@ -387,7 +353,7 @@ async def authenticate_google_user(request: GoogleLoginRequest,db:db_dependency)
         "role":"user"
     }
 
-    existing_employee = db.query(models.EmployeeDetails).filter_by(employee_id=user_data["employee_id"]).first()
+    existing_employee = db.query(models.EmployeeDetails).filter(models.EmployeeDetails.employee_id==user_data["employee_id"]).first()
 
     if existing_employee:
         pass
